+++ cardano-config/src/Cardano/Config/Protocol/Byron.hs
@@ -37,7 +37,7 @@ import           Cardano.Config.Types
                     DelegationCertFile (..), SigningKeyFile (..),
                     Update (..), LastKnownBlockVersion (..))
 import           Cardano.Config.Protocol.Types (SomeConsensusProtocol(..))
-import           Cardano.TracingInstances.Byron ()
+import           Cardano.TracingOrphanInstances.Byron ()
 
 
 ------------------------------------------------------------------------------
